refactor Main js, Calendar css
