+++ src/App.css
@@ -75,27 +75,6 @@
   color: gray;
 }
 
-/* calendar */
-.fc-theme-standard {
-  width: 100%;
-  height: 85vh;
-  color: var(--dark-green);
-}
-
-.fc-toolbar {
-  display: none;
-}
-
-.fc-event-title {
-  word-wrap: break-word;
-}
-
-.fc-event-main-frame {
-  padding: 5px;
-}
-
-/* end of calendar */
-
 /* TODO: make this more real */
 .shadow {
   -moz-box-shadow:    3px 3px 2px 3px #ccc;
