+++ src/Calendar.css
@@ -1,0 +1,22 @@
+.fc-theme-standard {
+  width: 100%;
+  height: 95vh;
+  color: var(--dark-green);
+}
+
+.fc-header-toolbar {
+  display: none !important;
+}
+
+.fc-event-title {
+  word-wrap: break-word;
+}
+
+.fc-event-main-frame {
+  padding: 5px;
+}
+
+.fc-col-header-cell-cushion {
+  color: var(--dark-green);
+  text-decoration: none;
+}
